feat: add cron function to update spotprice
